[refactor]#20 변수명 변경(CustomBottomTab -> BottomTab)
